change pointer cast type from int to intptr_t

+++ doc_msgs.c
@@ -338,7 +338,7 @@ int i;
     my_printf ("readmessage:  no mh!\n");
     return (MNFERR);
   }
-  if ((int)p & 3) // This is appears to be testing for an address not aligned to 32-bit boundary. But why?
+  if ((intptr_t)p & 3) // This is appears to be testing for an address not aligned to 32-bit boundary. But why?
   {
     printf ("wtf?\r\n");
     return(MNFERR);
